Refactor outputs module
Move data types to their respective modules
