+++ parley/src/outputs/alignment.rs
@@ -1,8 +1,19 @@
 // Copyright 2024 the Parley Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use super::{Alignment, BreakReason, LayoutData};
 use crate::inputs::style::Brush;
+use crate::outputs::{BreakReason, LayoutData};
+
+/// Alignment of a layout.
+#[derive(Copy, Clone, Default, PartialEq, Eq, Debug)]
+#[repr(u8)]
+pub enum Alignment {
+    #[default]
+    Start,
+    Middle,
+    End,
+    Justified,
+}
 
 pub(crate) fn align<B: Brush>(
     layout: &mut LayoutData<B>,
